# docs/swagger.yaml
basePath: /v1
definitions:
...
  auth.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
  auth.RefreshDTO:
    properties:
      refresh_token:
        minLength: 1
        type: string
      telegram_id:
        minLength: 1
        type: string
    required:
    - refresh_token
    - telegram_id
    type: object
  auth.RefreshResponse:
    properties:
      access_exp_at:
        type: string
      access_token:
        type: string
      refresh_exp_at:
        type: string
      refresh_token:
        type: string
    type: object
  auth.SignInDTO:
    properties:
      first_name:
        minLength: 1
        type: string
      last_name:
        minLength: 1
        type: string
      telegram_id:
        minLength: 1
        type: string
      username:
        minLength: 1
        type: string
    required:
    - first_name
    - telegram_id
    type: object
  auth.SignInResp:
    properties:
      access_exp_at:
        type: string
      access_token:
        type: string
      refresh_exp_at:
        type: string
      refresh_token:
        type: string
    type: object
...
  product.AddExcludeProductsByTelegramIDDTO:
    properties:
      products:
        items:
          type: string
        maxItems: 50
        minItems: 1
        type: array
      telegram_id:
        minLength: 1
        type: string
    required:
    - products
    - telegram_id
    type: object
  product.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
  product.UserExcludedProducts:
    properties:
      created_at:
        type: string
      id:
        type: integer
      products:
        items:
          type: string
        type: array
      telegram_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  recipe.Content:
    properties:
      bzhu:
        type: string
      calories:
        type: string
      id:
        type: integer
      ingredients:
        items:
          type: string
        type: array
      method_preparation:
        items:
          type: string
        type: array
      recipe_preparation:
        type: string
      title:
        type: string
      type:
        type: string
    type: object
  recipe.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
...
  recipe.Recipes:
    properties:
      content:
        items:
          items:
            $ref: '#/definitions/recipe.Content'
          type: array
        type: array
      created_at:
        type: string
      id:
        type: integer
      telegram_id:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
...
  user.ErrorResponse:
    properties:
      detail:
        type: string
      error:
        type: string
    type: object
  user.ExistsDTO:
    properties:
      telegram_id:
        minLength: 1
        type: string
      username:
        minLength: 1
        type: string
    required:
    - telegram_id
    type: object
...
  user.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      telegram_id:
        type: string
      updated_at:
        type: string
      username:
        type: string
    type: object
host: localhost:50050
info:
  contact: {}
  description: This is a application for telegram web apps application.
  title: API application for telegram web apps application foodgrammm
  version: "1.0"
paths:
...
  /v1/auth/refresh:
    post:
      consumes:
      - application/json
      description: Refresh the access token using the provided Telegram ID and refresh
        token
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Refresh request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.RefreshDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with new tokens
          schema:
            $ref: '#/definitions/auth.RefreshResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Refresh user token
      tags:
      - Authentication
  /v1/auth/signin:
    post:
      consumes:
      - application/json
      description: Sign in a user using their Telegram ID, username, first name, and
        last name
      parameters:
      - description: Sign in request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/auth.SignInDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Successful response with tokens
          schema:
            $ref: '#/definitions/auth.SignInResp'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/auth.ErrorResponse'
      summary: Sign in user
      tags:
      - Authentication
...
  /v1/product/exclude/telegram/{telegramID}:
    delete:
      consumes:
      - application/json
      description: This endpoint allows you to delete excluded products for a user
        identified by their Telegram ID.
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Telegram ID of the user
        in: path
        name: telegramID
        required: true
        type: string
      - description: Product name to be deleted
        in: query
        name: product
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully deleted excluded product
          schema:
            $ref: '#/definitions/product.UserExcludedProducts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/product.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/product.ErrorResponse'
      summary: Delete excluded products by Telegram ID
      tags:
      - Product
    get:
      consumes:
      - application/json
      description: Get excluded products for a user by their unique Telegram ID.
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Telegram ID
        in: path
        name: telegramID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Excluded products for the user
          schema:
            $ref: '#/definitions/product.UserExcludedProducts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/product.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/product.ErrorResponse'
      summary: Get Exclude Products by Telegram ID
      tags:
      - Product
  /v1/product/exclude/telegram/id:
    post:
      consumes:
      - application/json
      description: Add excluded products for a user by their unique Telegram ID.
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Exclude products data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/product.AddExcludeProductsByTelegramIDDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Excluded products for the user
          schema:
            $ref: '#/definitions/product.UserExcludedProducts'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/product.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/product.ErrorResponse'
      summary: Add Exclude Products by Telegram ID
      tags:
      - Product
...
  /v1/recipe/telegram/{telegramID}:
    get:
      consumes:
      - application/json
      description: Retrieve recipes for a user by their Telegram ID
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Telegram ID of the user
        in: path
        name: telegramID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of recipes
          schema:
            items:
              $ref: '#/definitions/recipe.Recipes'
            type: array
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/recipe.ErrorResponse'
        "404":
          description: User not found error
          schema:
            $ref: '#/definitions/recipe.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/recipe.ErrorResponse'
      summary: Get Recipes by Telegram ID
      tags:
      - Recipe
...
  /v1/user/all:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all users.
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/user.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Get all users
      tags:
      - User
  /v1/user/exists:
    post:
      consumes:
      - application/json
      description: Check if a user exists by Telegram ID and Username.
      parameters:
      - description: User data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/user.ExistsDTO'
      produces:
      - application/json
      responses:
        "200":
          description: Boolean flag indicating if the user exists
          schema:
            type: boolean
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Check if a user exists
      tags:
      - User
...
  /v1/user/telegram/{telegramID}:
    delete:
      consumes:
      - application/json
      description: Delete a user by their unique Telegram ID.
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Telegram ID
        in: path
        name: telegramID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Deleted Telegram ID
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Delete a user by Telegram ID
      tags:
      - User
    get:
      consumes:
      - application/json
      description: Retrieve a user by their unique Telegram ID.
      parameters:
      - default: Bearer <token>
        description: Authorization token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Telegram ID
        in: path
        name: telegramID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User details
          schema:
            $ref: '#/definitions/user.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/user.ErrorResponse'
      summary: Get a user by Telegram ID
      tags:
      - User
...
swagger: "2.0"